Rename qibolab.utils to qibolab.serialize
