+++ src/qibolab/dummy.py
@@ -1,7 +1,7 @@
 from qibolab.channels import Channel, ChannelMap
 from qibolab.instruments.dummy import DummyInstrument
 from qibolab.platform import Platform
-from qibolab.utils import load_qubits, load_settings
+from qibolab.serialize import load_qubits, load_settings
 
 NAME = "dummy"
 RUNCARD = {
