+++ tests/dummy_qrc/qblox.py
@@ -26,7 +26,7 @@ from qibolab.instruments.qblox.port import (
 )
 from qibolab.instruments.rohde_schwarz import SGS100A
 from qibolab.platform import Platform
-from qibolab.utils import load_qubits, load_runcard, load_settings
+from qibolab.serialize import load_qubits, load_runcard, load_settings
 
 NAME = "qblox"
 ADDRESS = "192.168.0.6"
